+++ src/main/kotlin/land/tbp/y2019/intcode/computer/instructions/MultiplyInstruction.kt
@@ -13,7 +13,7 @@ object MultiplyInstruction : Instruction {
         return values[0] * values[1]
     }
 
-    override fun size(): Byte {
+    override fun size(): Int {
         return 4
     }
 }
